pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pdfplumber>=0.11.7",
    "psycopg2-binary>=2.9.10",
    "pytesseract>=0.3.13",
    "pytest>=8.4.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "flask-migrate>=4.1.0",
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
]
